+++ com.unity.cinemachine/Editor/Editors/CinemachineConfiner2DEditor.cs
@@ -34,11 +34,10 @@ namespace Cinemachine.Editor
             var volumeProp = serializedObject.FindProperty(() => Target.BoundingShape2D);
             ux.Add(new PropertyField(volumeProp));
             ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.Damping)));
-            
+            ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.SlowingDistance)));
             var oversizedCameraHelp = ux.AddChild(new HelpBox(
                 "The camera window is too big for the confiner. Enable the Oversize Window option.",
                 HelpBoxMessageType.Info));
-            
             ux.Add(new PropertyField(serializedObject.FindProperty(() => Target.OversizeWindow)));
             
             TrackVolume(volumeProp);