+++ com.unity.cinemachine/Editor/Editors/CinemachineStateDrivenCameraEditor.cs
@@ -105,6 +105,8 @@ namespace Cinemachine.Editor
             EditorGUILayout.Separator();
             if (m_ChildListHelper.OnInspectorGUI(FindProperty(x => x.m_ChildCameras)))
                 Target.ValidateInstructions();
+            if (EditorGUI.EndChangeCheck()) 
+                serializedObject.ApplyModifiedProperties();
 
             // Extensions
             DrawExtensionsWidgetInInspector();