+++ com.unity.cinemachine/Editor/PropertyDrawers/SplineAutoDollyPropertyDrawer.cs
@@ -8,14 +8,12 @@ using System.Reflection;
 
 namespace Cinemachine.Editor
 {
-    [CustomPropertyDrawer(typeof(SplineAutoDolly))]
+    [CustomPropertyDrawer(typeof(SplineAutoDolly.ISplineAutoDolly), true)]
     class SplineAutoDollyPropertyDrawer : PropertyDrawer
     {
-        SerializedProperty GetImplementation(SerializedProperty p) => p.FindPropertyRelative("Implementation");
-
         int GetImplementationIndex(SerializedProperty p)
         {
-            var value = GetImplementation(p).managedReferenceValue;
+            var value = p.managedReferenceValue;
             return AutoDollyMenuItems.GetTypeIndex(value == null ? null : value.GetType());
         }
 
@@ -36,12 +34,11 @@ namespace Cinemachine.Editor
                 var index = AutoDollyMenuItems.GetTypeIndex(evt.newValue);
                 if (index != GetImplementationIndex(property))
                 {
-                    var p = GetImplementation(property);
-                    var targets = p.serializedObject.targetObjects;
+                    var targets = property.serializedObject.targetObjects;
                     foreach (var t in targets)
                     {
                         var o = new SerializedObject(t);
-                        var p2 = o.FindProperty(p.propertyPath);
+                        var p2 = o.FindProperty(property.propertyPath);
                         p2.managedReferenceValue = (index == 0) 
                             ? null : Activator.CreateInstance(AutoDollyMenuItems.s_AllItems[index]);
                         o.ApplyModifiedProperties();
@@ -72,7 +69,7 @@ namespace Cinemachine.Editor
                 {
                     var foldout = ux.AddChild(new Foldout() { name = kElementName, value = true });
                     foldout.Q<Toggle>(className: "unity-foldout__toggle").RemoveFromHierarchy(); // remove the header
-                    var childProperty = GetImplementation(property);
+                    var childProperty = property.Copy();
                     var endProperty = childProperty.GetEndProperty();
                     childProperty.NextVisible(true);
                     while (!SerializedProperty.EqualContents(childProperty, endProperty))