+++ com.unity.cinemachine/Runtime/Behaviours/CinemachineBrain.cs
@@ -118,11 +118,11 @@ namespace Cinemachine
         [FormerlySerializedAs("m_WorldUpOverride")]
         public Transform WorldUpOverride;
 
-        /// <summary>The CinemachineBrain will find the highest-priority CmCamera that outputs to any of the channels selected. 
-        /// CmCameras that do not output to one of these channels will be ignored.  Use this in situations where multiple
+        /// <summary>The CinemachineBrain will find the highest-priority CinemachineCamera that outputs to any of the channels selected. 
+        /// CinemachineCameras that do not output to one of these channels will be ignored.  Use this in situations where multiple
         /// CinemachineBrains are needed (for example, Split-screen).</summary>
-        [Tooltip("The CinemachineBrain will find the highest-priority CmCamera that outputs to any of the channels selected. "
-            + "CmCameras that do not output to one of these channels will be ignored.  Use this in situations "
+        [Tooltip("The CinemachineBrain will find the highest-priority CinemachineCamera that outputs to any of the channels selected. "
+            + "CinemachineCameras that do not output to one of these channels will be ignored.  Use this in situations "
             + "where multiple CinemachineBrains are needed (for example, Split-screen).")]
         [EnumMaskProperty]
         public OutputChannel.Channels ChannelMask = OutputChannel.Channels.Default;