+++ com.unity.cinemachine/Runtime/Deprecated/AxisState.cs
@@ -165,7 +165,7 @@ namespace Cinemachine
 
         /// <summary>
         /// This is an interface to override default querying of Unity's legacy Input system.
-        /// If a befaviour implementing this interface is attached to a CmCamera that 
+        /// If a befaviour implementing this interface is attached to a CinemachineCamera that 
         /// requires input, that interface will be polled for input instead of the standard Input system.
         /// </summary>
         [Obsolete("IInputAxisProvider is deprecated.  Use InputAxis and InputAxisController instead")]
@@ -221,7 +221,7 @@ namespace Cinemachine
             m_LastUpdateFrame = Time.frameCount;
 
             // Cheating: we want the render frame time, not the fixed frame time
-            if (deltaTime >= 0 && m_LastUpdateTime != 0) 
+            if (deltaTime > 0 && m_LastUpdateTime != 0) 
                 deltaTime = Time.realtimeSinceStartup - m_LastUpdateTime;
             
             m_LastUpdateTime = Time.realtimeSinceStartup;
@@ -437,7 +437,7 @@ namespace Cinemachine
             public void DoRecentering(ref AxisState axis, float deltaTime, float recenterTarget)
             {
                 // Cheating: we want the render frame time, not the fixed frame time
-                if (deltaTime >= 0)
+                if (deltaTime > 0)
                     deltaTime = Time.realtimeSinceStartup - m_LastUpdateTime;
                 
                 m_LastUpdateTime = Time.realtimeSinceStartup;