+++ com.unity.cinemachine/Tests/Runtime/ClearShotTests.cs
@@ -12,8 +12,8 @@ namespace Tests.Runtime
     {
         GameObject m_Character;
         CinemachineClearShot m_ClearShot;
-        CmCamera m_Vcam1;
-        CmCamera m_Vcam2;
+        CinemachineCamera m_Vcam1;
+        CinemachineCamera m_Vcam2;
         
         [SetUp]
         public override void SetUp()
@@ -32,18 +32,18 @@ namespace Tests.Runtime
             clearShotCollider.MinimumDistanceFromTarget = 0.1f;
 
             // a stationary vcam1 with a hard lookat
-            var vcam1Holder = CreateGameObject("CM Vcam1", typeof(CmCamera));
+            var vcam1Holder = CreateGameObject("CM Vcam1", typeof(CinemachineCamera));
             vcam1Holder.transform.SetParent(clearShotHolder.transform);
             vcam1Holder.transform.position = new Vector3(0, 0, 8); 
-            m_Vcam1 = vcam1Holder.GetComponent<CmCamera>();
+            m_Vcam1 = vcam1Holder.GetComponent<CinemachineCamera>();
             m_Vcam1.gameObject.AddComponent<CinemachineHardLookAt>();
             m_Vcam1.Priority = 20;
             
             // a completely locked vcam2
-            var vcam2Holder = CreateGameObject("CM Vcam2", typeof(CmCamera));
+            var vcam2Holder = CreateGameObject("CM Vcam2", typeof(CinemachineCamera));
             vcam2Holder.transform.SetParent(clearShotHolder.transform);
             vcam2Holder.transform.position = new Vector3(0, 0, -2);
-            m_Vcam2 = vcam2Holder.GetComponent<CmCamera>();
+            m_Vcam2 = vcam2Holder.GetComponent<CinemachineCamera>();
             m_Vcam2.Priority = 10;
 
             // a "wall" composed of a single quad that partially obscures vcam1, but not vcam2